Serialize Slack notifications in follow-up job
