+++ .github/scripts/pull-request-dashboard/utils.py
@@ -1,0 +1,56 @@
+from __future__ import annotations
+
+from datetime import datetime, timezone
+from typing import Any
+
+
+DEFAULT_TRUNCATE_CHARS = 1200
+
+
+def parse_ts(s: str | None) -> datetime | None:
+    if not s:
+        return None
+    try:
+        return datetime.fromisoformat(s.replace("Z", "+00:00"))
+    except ValueError:
+        return None
+
+
+def seconds_since(ts: datetime | None) -> int | None:
+    if ts is None:
+        return None
+    return max(0, int((datetime.now(timezone.utc) - ts).total_seconds()))
+
+
+def activity_age(ts: datetime | None) -> str:
+    seconds = seconds_since(ts)
+    if seconds is None:
+        return "?"
+    minutes = seconds // 60
+    if minutes < 1:
+        return "<1m"
+    if minutes < 60:
+        return f"{minutes}m"
+    hours = minutes // 60
+    if hours < 24:
+        return f"{hours}h"
+    return f"{hours // 24}d"
+
+
+def truncate(s: str, n: int = DEFAULT_TRUNCATE_CHARS) -> str:
+    s = (s or "").strip()
+    if len(s) <= n:
+        return s
+    return s[:n] + " ...[truncated]"
+
+
+def actor_login(obj: dict[str, Any] | None) -> str:
+    return ((obj or {}).get("login") or "").strip()
+
+
+def format_ts(ts: datetime | None) -> str:
+    return ts.isoformat() if ts else ""
+
+
+def utc_now() -> datetime:
+    return datetime.now(timezone.utc)
